pyproject.toml
[project]
name = "scholarly-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.17",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.33",
    "langchain-openai>=0.3.3",
    "semanticscholar>=0.9.0",
    "ipykernel>=6.29.5",
    "xmltodict>=0.14.2",
    "langgraph>=0.2.69",
    "python-dotenv>=1.0.1",
    "wat>=0.5.0",
    "tabulate>=0.9.0",
    "more-itertools>=10.6.0",
    "langchain-google-genai>=2.0.9",
    "ipywidgets>=8.1.5",
    "tomli>=2.2.1",
    "metapub>=0.5.12",
]
